MapToSlice in common
